+++ inc/HM11/hm11.c
@@ -48,7 +48,7 @@ void Hm11_Packet(void){
 //	_tx_buffer_[19] = 0x0A;
 
     /* minimum delay before UART transmission */
-	for(int i=0; i<250; i++){for(int j=0; j<250; j++){;}}
+//	for(int i=0; i<250; i++){for(int j=0; j<250; j++){;}}
 }
 
 void FabricIrq1_IRQHandler(void){
@@ -64,8 +64,8 @@ void SysTick_Handler(void) {
     static uint32_t count = 0;
     I2C_system_tick(&g_core_i2c0, 1);
     if(_tx_buffer_ != ( int8_t* ) 0){
-        if (count == 5000){
-            // UART tx every 5000 ticks (10ms)
+        if (count == 10){
+            // UART tx every 10 ticks (10ms)
     		Hm11_Packet();
             UART_send( &g_uart_0, _tx_buffer_, 20 );
             count = 0;