+++ inc/MPU6050/mpu6050.h
@@ -27,7 +27,7 @@
 /* ----- I2C Related ----- */
 #define COREI2C_BASE_ADDR   0x50000000
 #define COREI2C_SER_ADDR    0xFF
-#define I2C_TIMEOUT         10        // ms
+#define I2C_TIMEOUT         1        // ms
 
 /* ----- MPU6050 Related ----- */
 #define MPU6050_ADDR_LOW    0x68    // 7-Bit address when AD0 is connect to GND //